Revert "Attainment Fix"
