+++ Header.php
@@ -9,7 +9,7 @@
         <link rel="stylesheet" href="https://www.w3schools.com/w3css/4/w3.css">
         <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lato">
         <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Montserrat">
-        <!--<link rel="stylesheet" href="https://www.w3schools.com/lib/w3-theme-black.css">-->
+        <link rel="stylesheet" href="https://www.w3schools.com/lib/w3-theme-black.css">
         <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">  
         <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
         <link href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.0.3/css/bootstrap.min.css" rel="stylesheet" type="text/css" />
@@ -55,9 +55,5 @@
                 top: 60px;  /* chartArea top  */
                 left: 180px; /* chartArea left */
             }
-            
- select.selector{
-color:red;
-}
         </style>
