--- css/print.css
@@ -1,55 +1,0 @@
-
-@font-face {
-    font-family: monospace;
-    /*src: url(sansation_light.woff);*/
-}
-
-
-.selectprint { padding-right:  0px;
-  border: 0;
-  margin-bottom:5px;
-  outline: none;
-
-}
-
-table {
-  border-collapse: collapse;
- 
-}
-
-table,  td {
-  border: 1px solid black;
-}
-
-th, td {
-  padding: 15px;
-  
-}
-input, select {
-    max-width: 100px;
-    border: none;
-}
-
-p { font-size: 20px;}
-
-
-
-.h5 {font-size: 30px;}
-
-
-
-select {
-  /* for Firefox */
-  -moz-appearance: none;
-  /* for Chrome */
-  -webkit-appearance: none;
-}
-
-/* For IE10 */
-select::-ms-expand {
-  display: none;
-}
-
-.arabictd{
-    text-align: right;
-}
