+++ css/resultstyle.css
@@ -1,25 +1,5 @@
-.wrapper{
-    border: 1px dashed red;
-    height: 150px;
-    overflow-x: hidden;
-    overflow-y: scroll;
-    width: 150px;
- }
- .wrapper .selection{
-   width:150px;
-   border:1px solid #ccc
- }
-.loader_div{
-  position: absolute;
-  top: 0;
-  bottom: 0%;
-  left: 0;
-  right: 0%;
-  z-index: 99;
-  opacity:0.7;
-  display:none;
-  background: lightgrey url('imges/Preloader_2.gif') center center no-repeat;
-}
+
+
 input[type=text]{
     width: 25%;
     font-size: 11px;
@@ -116,6 +96,9 @@ input[type=text]{
 }
 
 
+/* Paste this css to your style sheet file or under head tag */
+/* This only works with JavaScript, 
+if it's not present, don't show loader */
 .no-js #loader { display: none;  }
 .js #loader { display: block; position: absolute; left: 100px; top: 0; }
 .se-pre-con {
