--- include/popContainer.php
@@ -1,109 +1,0 @@
-
-<div id = "popcontainer" class="popover-content-el hide " style="width:400px; "  >
-
-    <div id="chart_div" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left; cursor: pointer; color: gray">Click to view details</h6>
-    <button class="  w3-hover-teal  w3-round-xxlarge " id="exportS" style="float: right; margin-bottom: 10px; color: teal" onclick="downloadPopoverStatistics()" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
-
-<div id = "popcontainerSubject1" class="popover-content-el hide  " style="width:400px; "  >
-    <div id="chart1" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left;cursor: pointer; color: gray">Click to view details</h6>
-    <button class="  w3-hover-teal  w3-round-xxlarge "  style="float: right; margin-bottom: 10px; color: teal" onclick="downloadPopoverSubjects(1)" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
-
-<div id = "popcontainerSubject2" class="popover-content-el hide " style="width:400px; "  >
-    <div id="chart2" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left;cursor: pointer; color: gray">Click to view details</h6>
-    <button class="  w3-hover-teal  w3-round-xxlarge "  style="float: right; margin-bottom: 10px; color: teal" onclick="downloadPopoverSubjects(2)" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
-
-<div id = "popcontainerSubject3" class="popover-content-el hide " style="width:400px; "  >
-    <div id="chart3" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left;cursor: pointer; color: gray">Click to view details</h6>
-    <button class=" w3-hover-teal w3-round-xxlarge "  style="float: right; color: teal; margin-bottom: 10px" onclick="downloadPopoverSubjects(3)" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
-<div id = "popcontainerSubject4" class="popover-content-el hide " style="width:400px; "  >
-    <div id="chart4" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left; cursor: pointer; color: gray">Click to view details</h6>
-    <button class="w3-hover-teal w3-round-xxlarge"  style="float: right;color : teal; margin-bottom: 10px" onclick="downloadPopoverSubjects(4)" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
-<div id = "popcontainerSubject5" class="popover-content-el hide  " style="width:400px; "  >
-    <div id="chart5" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left;cursor: pointer; color: gray">Click to view details</h6>
-    <button class="  w3-hover-teal w3-round-xxlarge "  style="float: right; margin-bottom: 10px; color :teal" onclick="downloadPopoverSubjects(5)" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
-<div id = "popcontainerSubject6" class="popover-content-el hide " style="width:400px; "  >
-    <div id="chart6" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left;cursor: pointer; color: gray">Click to view details</h6>
-    <button class="w3-hover-teal  w3-round-xxlarge "  style="float: right;color :teal; margin-bottom: 10px" onclick="downloadPopoverSubjects(6)" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
-<div id = "popcontainerSubject7" class="popover-content-el hide " style="width:400px; "  >
-    <div id="chart7" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left;cursor: pointer; color: gray">Click to view details</h6>
-    <button class="  w3-hover-teal  w3-round-xxlarge "  style="float: right; margin-bottom: 10px; color : teal" onclick="downloadPopoverSubjects(7)" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
-<div id = "popcontainerSubject8" class="popover-content-el hide " style="width:400px; "  >
-    <div id="chart8" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left;cursor: pointer; color: gray">Click to view details</h6>
-    <button class=" w3-hover-teal  w3-round-xxlarge "  style="float: right; margin-bottom: 10px; color : teal" onclick="downloadPopoverSubjects(8)" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
-<div id = "popcontainerSubject9" class="popover-content-el hide " style="width:400px; "  >
-    <div id="chart9" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left;cursor: pointer; color: gray">Click to view details</h6>
-    <button class="  w3-hover-teal w3-round-xxlarge "  style="float: right; margin-bottom: 10px; color:teal" onclick="downloadPopoverSubjects(9)" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
-<div id = "popcontainerSubject10" class="popover-content-el hide " style="width:400px; "  >
-    <div id="chart10" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left;cursor: por-blue-grayinter; color: gray">Click to view details</h6>
-    <button class="  w3-hover-teal w3-round-xxlarge "  style="float: right; color:teal; margin-bottom: 10px" onclick="downloadPopoverSubjects(10)" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
-<div id = "popcontainerSubject11" class="popover-content-el hide " style="width:400px; "  >
-    <div id="chart11" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left;cursor: pointer; color: gray">Click to view details</h6>
-    <button class="  w3-hover-teal w3-round-xxlarge "  style="float: right; margin-bottom: 10px; color : teal" onclick="downloadPopoverSubjects(11)" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
-<div id = "popcontainerSubject12" class="popover-content-el hide " style="width:400px; "  >
-    <div id="chart12" style="width:400px; "  >
-
-    </div>
-    <h6   style="float: left;cursor: pointer; color: gray">Click to view details</h6>
-    <button class=" w3-hover-teal  w3-round-xxlarge "  style="float: right;color : teal; margin-bottom: 10px" onclick="downloadPopoverSubjects(12)" title="Download Graph">
-        <span class="material-icons">save_alt</span></button>
-</div>
