--- javaScripts/loaderInitializer.js
@@ -1,5 +1,0 @@
-$(window).load(function () {
-    // Animate loader off screen
-    $(".se-pre-con").fadeOut("slow");
-    ;
-});
