--- javaScripts/popover.js
@@ -1,287 +1,0 @@
-
-
-$(document).ready(function () {
-    $('[data-toggle="popover"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainer').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
-
-
-
-$(document).ready(function () {
-    $('[data-toggle="popoverSubject1"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainerSubject1').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
-
-
-
-
-$(document).ready(function () {
-    $('[data-toggle="popoverSubject2"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainerSubject2').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
-
-
-
-$(document).ready(function () {
-    $('[data-toggle="popoverSubject3"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainerSubject3').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
-
-
-
-$(document).ready(function () {
-    $('[data-toggle="popoverSubject4"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainerSubject4').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
-
-
-
-$(document).ready(function () {
-    $('[data-toggle="popoverSubject5"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainerSubject5').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
-
-
-
-$(document).ready(function () {
-    $('[data-toggle="popoverSubject6"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainerSubject6').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
-
-
-
-$(document).ready(function () {
-    $('[data-toggle="popoverSubject7"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainerSubject7').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
-
-
-
-$(document).ready(function () {
-    $('[data-toggle="popoverSubject8"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainerSubject8').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
-
-
-
-$(document).ready(function () {
-    $('[data-toggle="popoverSubject9"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainerSubject9').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
-
-
-
-
-$(document).ready(function () {
-    $('[data-toggle="popoverSubject10"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainerSubject10').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
-
-
-
-$(document).ready(function () {
-    $('[data-toggle="popoverSubject11"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainerSubject11').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
-
-
-
-$(document).ready(function () {
-    $('[data-toggle="popoverSubject12"]').popover(
-            {
-                trigger: "manual",
-                html: true,
-                content: function () {
-
-                    return $('#popcontainerSubject12').html();
-                }
-            }).on("mouseenter", function () {
-        var _this = this;
-        $(this).popover("show");
-        $(".popover").on("mouseleave", function () {
-            $(_this).popover('hide');
-        });
-    });
-});
-
-
