+++ nbproject/project.properties
@@ -1,4 +1,3 @@
-browser.reload.on.save=true
 include.path=${php.global.include.path}
 php.version=PHP_73
 source.encoding=UTF-8
